Fix some test pyccel tests

+++ tests/internal/scripts/mpi/bcast.py
@@ -27,9 +27,9 @@ if __name__ == '__main__':
 
     master = np.int32(1)
     if rank == master:
-        msg = rank + 1000
+        msg = rank + np.int32(1000)
     else:
-        msg = 0
+        msg = np.int32(0)
 
     length = np.int32(1)
     mpi_bcast (msg, length, MPI_INTEGER8, master, comm, ierr)
